dora_node_hub/dora_andino_gemini_control/pyproject.toml
[project]
name = "dora_andino_gemini_control"
version = "0.1.0"
authors = [{ name = "Your Name", email = "email@email.com" }]
description = "dora_andino_gemini_control"
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.9"

dependencies = [
  "dora-rs >= 0.3.9",
  "google-genai >= 1.23.0",
  "numpy >= 1.26.4",
]

[dependency-groups]
dev = [
  "pytest >=8.1.1",
  "ruff >=0.9.1",
]

[project.scripts]
dora_andino_gemini_control = "dora_andino_gemini_control.main:main"

[tool.ruff.lint]
extend-select = [
  "D",   # pydocstyle
  "UP"
]
